% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\alias{SE}
\title{Example dataset}
\value{
a \code{\link[SummarizedExperiment]{SummarizedExperiment}}.
}
\description{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} containing (a
subset of) whole-hippocampus RNAseq of mice after different stressors.
}
\references{
Floriou-Servou et al. (2018).
Distinct Proteomic, Transcriptomic, and Epigenetic Stress Responses in Dorsal
 and Ventral Hippocampus.
\emph{Biological Psychiatry}, \bold{84}(7): 531-541.
DOI: 10.1016/j.biopsych.2018.02.003.
}
