% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltSE.R
\name{meltSE}
\alias{meltSE}
\title{meltSE}
\usage{
meltSE(x, genes, assayName = NULL, colDat.columns = NULL,
  rowDat.columns = NA)
}
\arguments{
\item{x}{An object of class
\code{\link[SummarizedExperiment]{SummarizedExperiment}}}

\item{genes}{A vector of genes to include. Use `genes=NULL` to include all.}

\item{assayName}{The name(s) of the assay(s) to use. If NULL and the assays are named,
all of them will be included (if they are not named, the first one will be used).}

\item{colDat.columns}{The colData columns to include (defaults includes all).
Use `colDat.columns=NA` in order not to include any.}

\item{rowDat.columns}{The rowData columns to include (none included by default). Use
`rowData=NULL` to include all.}
}
\value{
A data.frame.
}
\description{
Melts a SE object into a \code{\link[ggplot2]{ggplot}}-ready long data.frame.
}
\examples{
data("SE", package="SEtools")
head(meltSE(SE,"Fos"))

}
