% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{resetAllSEtoolsOptions}
\alias{resetAllSEtoolsOptions}
\title{resetAllSEtoolsOptions}
\usage{
resetAllSEtoolsOptions()
}
\value{
None
}
\description{
Resents all global options relative to SEtools.
}
\examples{
resetAllSEtoolsOptions()

}
