% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{se2xls}
\alias{se2xls}
\title{se2xlsx}
\usage{
se2xls(se, filename, addSheets = NULL)
}
\arguments{
\item{se}{The `SummarizedExperiment`}

\item{filename}{xlsx file name}

\item{addSheets}{An optional list of additional tables to save as sheets.}
}
\value{
Saves to file.
}
\description{
Writes a SummarizedExperiment to an excel/xlsx file. Requires the `openxlsx`
package.
}
