% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sehm.R
\name{sortRows}
\alias{sortRows}
\title{sortRows}
\usage{
sortRows(x, z = F, toporder = NULL, na.rm = F,
  method = "MDS_angle", toporder.meth = "before")
}
\arguments{
\item{x}{A numeric matrix or data.frame.}

\item{z}{Whether to scale rows for the purpose of calculating order (default FALSE).}

\item{toporder}{Optional verctor of categories (length=nrow(x)) on which to supra-order
when sorting rows.}

\item{na.rm}{Wheter to remove missing values and invariant rows (default FALSE).}

\item{method}{Seriation method; 'MDS_angle' (default) or 'R2E' recommended.}

\item{toporder.meth}{Whether to perform higher-order sorting 'before' (default) or 'after' the lower-
order sorting.}
}
\value{
A reordered matrix or data.frame.
}
\description{
sortRows
}
