% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinySEE.R
\name{tinySEE}
\alias{tinySEE}
\title{tinySEE}
\usage{
tinySEE(se = NULL, ...)
}
\arguments{
\item{se}{An optional object of class `SummarizedExperiment`. If omitted,
users will be able to upload one.}

\item{...}{Any parameters passed to `shinyApp`}
}
\value{
a shiny app
}
\description{
A very simple shiny app for exploring a SummarizedExperiment, allowing the
browsing and plotting of genes (essentially an interface to the SEtools
heatmaps). For more functionality, see the `iSEE` package.
}
\examples{
data(SE)
app <- tinySEE(SE)

}
