% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltSE.R
\name{meltSE}
\alias{meltSE}
\title{meltSE}
\usage{
meltSE(x, genes = NULL, assayName = NULL, colDat.columns = NULL,
  value.name = NULL)
}
\arguments{
\item{x}{An object of class `SummarizedExperiment`}

\item{genes}{A vector of genes to include}

\item{assayName}{The name of the assay to use. If NULL, the first one will be used.}

\item{colDat.columns}{The data columns to include (defaults includes all)}

\item{value.name}{The name of the value column.}
}
\value{
A data.frame.
}
\description{
Melts a SE object into a ggplot-compatible data.frame
}
