% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSEs.R
\name{mergeSEs}
\alias{mergeSEs}
\title{mergeSEs}
\usage{
mergeSEs(ll, what = "zscores", commonOnly = TRUE,
  assayName = c("logcpm", "lognorm"), colColumns = NULL,
  defValues = list())
}
\arguments{
\item{ll}{A (named) list of SummarizedExperiments}

\item{what}{Values to use, either "asis" (input values) or "zscores" (within-dataset row z-scores)}

\item{commonOnly}{Logical; whether to restrict to rows present in all datasets (default TRUE).}

\item{assayName}{An optional vector of assayNames to use. The first available will be used, or the first assay if NULL.}

\item{colColumns}{A character vector specifying `colData` columns to include (if available in at least one of the datasets).
If NULL, everything is kept.}

\item{defValues}{A list specifying the default values for `colColumns` when these are absent.}
}
\value{
An object of class `SummarizedExperiment`
}
\description{
Merges a list of SummarizedExperiments.
}
