% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sehm.R
\name{sehm}
\alias{sehm}
\title{sehm}
\usage{
sehm(se, genes = NULL, do.scale = FALSE,
  assayName = .getDef("assayName"), sortRowsOn = 1:ncol(se),
  cluster_cols = FALSE, cluster_rows = is.null(sortRowsOn),
  toporder = NULL, hmcols = NULL, breaks = .getDef("breaks"),
  gaps_at = .getDef("gaps_at"), anno_rows = .getDef("anno_rows"),
  anno_columns = .getDef("anno_columns"),
  anno_colors = .getDef("anno_colors"), show_rownames = NULL,
  show_colnames = FALSE, ...)
}
\arguments{
\item{se}{A \code{\link[SummarizedExperiment]{SummarizedExperiment}}.}

\item{genes}{An optional vector of genes (i.e. row names of `se`)}

\item{do.scale}{Logical; whether to scale rows (default FALSE).}

\item{assayName}{An optional vector of assayNames to use. The first available
will be used, or the first assay if NULL.}

\item{sortRowsOn}{Sort rows by MDS polar order using the specified columns
(default all)}

\item{cluster_cols}{Whether to cluster columns (default F)}

\item{cluster_rows}{Whether to cluster rows; default FALSE if
`do.sortRows=TRUE`.}

\item{toporder}{Optional verctor of categories on which to supra-order when
sorting rows.}

\item{hmcols}{Colors for the heatmap.}

\item{breaks}{Breaks for the heatmap colors. Alternatively, if
`breaks==TRUE`, a symmetrical scale with capped ends will be used
(appropriate when plotting log2 foldchanges)}

\item{gaps_at}{Columns of `colData` to use to establish gaps between columns.}

\item{anno_rows}{Columns of `rowData` to use for annotation.}

\item{anno_columns}{Columns of `colData` to use for annotation.}

\item{anno_colors}{List of colors to use for annotation.}

\item{show_rownames}{Whether to show row names (default TRUE if 50 rows or
less).}

\item{show_colnames}{Whether to show column names (default FALSE).}

\item{...}{Further arguments passed to `pheatmap`.}
}
\value{
A heatmap (see \code{\link[pheatmap]{pheatmap}}).
}
\description{
Heatmap wrapper for \code{\link[SummarizedExperiment]{SummarizedExperiment}}.
}
\examples{
data("SE", package="SEtools")
sehm(SE, do.scale=TRUE)

}
