% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svacor.R
\name{svacor}
\alias{svacor}
\title{svacor}
\usage{
svacor(
  SE,
  form,
  form0 = ~1,
  assayName = NULL,
  regressOutNull = TRUE,
  useVST = TRUE,
  n.sv = NULL,
  ...
)
}
\arguments{
\item{SE}{An object of class `SummarizedExperiment`.}

\item{form}{The formula of the differential expression model}

\item{form0}{An optional formula for the null model}

\item{assayName}{The name (or index) of the assay to use.}

\item{regressOutNull}{Logical; whether to regress out the variables of `form0`.}

\item{useVST}{Logical; whether to use DESeq2's variance-stabilizing transformation;
(for count data!)}

\item{n.sv}{The number of surrogate variables (if omitted, \code{\link{sva}} will
attempt to estimate it)}

\item{...}{Any other argument passed to the \code{\link{sva}} command.}
}
\value{
Returns the `SummarizedExperiment` with a `corrrected` assay and the surrogate
variables in `colData`.
}
\description{
A wrapper around SVA-based correction, providing a corrected assay. If this is RNAseq
data or similar, use a count assay assay with `useVST=TRUE`; otherwise (e.g.
proteomics) a log-normalized assay is recommended.
}
