% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svacor.R
\name{svacor}
\alias{svacor}
\title{svacor}
\usage{
svacor(
  SE,
  form = NULL,
  form0 = ~1,
  mm = NULL,
  mm0 = NULL,
  regressOutNull = TRUE,
  trans = c("svaseq", "none", "log"),
  assayName = NULL,
  n.sv = NULL,
  ...
)
}
\arguments{
\item{SE}{An object of class `SummarizedExperiment`. Alternatively, a matrix can be
used, but many options will not be supported.}

\item{form}{The formula of the differential expression model}

\item{form0}{An optional formula for the null model}

\item{mm}{If `form=NULL`, the model.matrix}

\item{mm0}{An optional null model.matrix.}

\item{regressOutNull}{Logical; whether to regress out the variables of `form0` (default TRUE)}

\item{trans}{Either "svaseq" (default, expects a count assay), "none", or "log".}

\item{assayName}{Which assay to use (if `SE` is a `SummarizedExperiment`); if missing,
will use the first one.}

\item{n.sv}{The number of surrogate variables (if omitted, `sva` will attempt to
estimate it).}

\item{...}{Any other argument passed to the `sva` command.}
}
\value{
Returns a `SummarizedExperiment` (with a `corrected` assay and the surrogate
variables in `colData`) if `SE` is an object of that class `SummarizedExperiment`;
otherwise, a list with the slots:
* `sv`: a table of the surrogate variables
* `cor`: the corrected data (for plotting)
* `mm`: the model.matrix containing, in addition to the specified experimental
variables, all detected surrogate variables.
}
\description{
A wrapper around SVA-based correction, providing a corrected assay.
}
